Update and improve HDF5 build (#1538)
* Change hdf5hl_fortran to hdf5_hl_fortran

This is the "correct" name; the make build creates a file with the wrong
name, and then, except on a Mac, creates a link to the right name. The
cmake build does the right thing. Requires a corresponding change to
bmad-external-packages.

* Get the right HDF5 libraries from find_package

* Use the cmake variable ACC_BMAD_LINK_LIBS

Stop using the environment variable in cmake files

* Revert "Use the cmake variable ACC_BMAD_LINK_LIBS"

This reverts commit ed5717d53f1258b812b94f81ad957ee4c455c00b.
Need to use environment variable since Master.cmake has not been called yet.

* Append HDF5 library to SHARED_LINK_LIBS

* Add HDF5_Fortran_LIBRARIES to SHARED_LINK_LIBS

On RHEL9, not automatically included as a dependency of
HDF5_Fortran_HL_LIBRARIES

---------

Co-authored-by: J. Scott Berg <jsberg@bnl.gov>